+++ packages/storybook/src/common/stories.ts
@@ -371,6 +371,21 @@ const LineString: Story = {
 };
 
 // Linestring drawing story
+const LineStringFinishOnNthCoordinate: Story = {
+	...DefaultStory,
+	args: {
+		id: "linestring-finish-on-nth-coordinate",
+		modes: [
+			() =>
+				new TerraDrawLineStringMode({
+					finishOnNthCoordinate: 2,
+				}),
+		],
+		...DefaultStory.args,
+	},
+};
+
+// Linestring drawing story
 const LineStringWithCoordinatePoints: Story = {
 	...DefaultStory,
 	args: {
@@ -792,6 +807,7 @@ const AllStories = {
 	AngledRectangle,
 	Sector,
 	LineString,
+	LineStringFinishOnNthCoordinate,
 	LineStringWithCoordinatePoints,
 	LineStringEditable,
 	LineStringWithCoordinateSnapping,